@@ -127,7 +127,7 @@
     "@electron/rebuild": "3.6.0",
     "@joplin/default-plugins": "~3.1",
     "@joplin/tools": "~3.1",
-    "@playwright/test": "1.44.1",
+    "@playwright/test": "1.45.3",
     "@testing-library/react-hooks": "8.0.1",
     "@types/jest": "29.5.12",
     "@types/node": "18.19.42",
@@ -137,7 +137,7 @@
     "@types/styled-components": "5.1.32",
     "@types/tesseract.js": "2.0.0",
     "axios": "^1.7.7",
-    "electron": "29.4.5",
+    "electron": "32.2.0",
     "electron-builder": "24.13.3",
     "glob": "10.4.5",
     "gulp": "4.0.2",
