@@ -14,20 +14,11 @@ export default function styles(themeId: number) {
 				display: 'flex',
 				flexDirection: 'row',
 			},
-			recorderContainer: {
-				padding: 2,
-				flexGrow: 1,
-			},
 			filterInput: {
 				...theme.inputStyle,
 				flexGrow: 1,
 				minHeight: 29,
 				alignSelf: 'center',
-			},
-			recorderInput: {
-				...theme.inputStyle,
-				minHeight: 29,
-				width: '200px',
 			},
 			label: {
 				...theme.textStyle,
@@ -48,10 +39,6 @@ export default function styles(themeId: number) {
 				...theme.textStyle,
 				width: 'auto',
 			},
-			tableCell: {
-				display: 'flex',
-				flexDirection: 'row',
-			},
 			tableCellContent: {
 				flexGrow: 1,
 				alignSelf: 'center',
@@ -59,6 +46,8 @@ export default function styles(themeId: number) {
 			tableCellStatus: {
 				height: '100%',
 				alignSelf: 'center',
+				border: 'none',
+				background: 'transparent',
 			},
 			kbd: {
 				fontFamily: 'sans-serif',
