@@ -1,37 +1,14 @@
-// Currently disabled on non-Android platforms
+import { VoiceTypingProvider } from './VoiceTyping';
 
-// eslint-disable-next-line @typescript-eslint/no-explicit-any -- Old code before rule was applied
-type Vosk = any;
-
-export { Vosk };
-
-interface StartOptions {
-	onResult: (text: string)=> void;
-}
-
-export interface Recorder {
-	stop: ()=> Promise<string>;
-	cleanup: ()=> void;
-}
-
-export const isSupportedLanguage = (_locale: string) => {
-	return false;
+const vosk: VoiceTypingProvider = {
+	supported: () => false,
+	modelLocalFilepath: () => null,
+	getDownloadUrl: () => null,
+	getUuidPath: () => null,
+	build: async () => {
+		throw new Error('Unsupported!');
+	},
+	modelName: 'vosk',
 };
 
-export const modelIsDownloaded = async (_locale: string) => {
-	return false;
-};
-
-export const getVosk = async (_locale: string) => {
-	// eslint-disable-next-line @typescript-eslint/no-explicit-any -- Old code before rule was applied
-	return {} as any;
-};
-
-export const startRecording = (_vosk: Vosk, _options: StartOptions): Recorder => {
-	return {
-		stop: async () => { return ''; },
-		cleanup: () => {},
-	};
-};
-
-export const voskEnabled = false;
+export default vosk;
